manually expose feature state methods
