--- projects/ngx-maplibre-gl/src/lib/source/source.ts
@@ -1,43 +1,0 @@
-import { FeatureIdentifier } from "maplibre-gl";
-import { SourceDirective } from "./source.directive";
-
-/** Abstract class for source components for shared methods. */
-export abstract class Source {
-  protected sourceDirective: SourceDirective;
-
-  /**
-   * Sets the `state` of a feature.
-   * 
-   * @param feature Feature identifier. `source` may be omitted, will use this source id.
-   * @param state A set of key-value pairs. The values should be valid JSON types.
-   * 
-   * @see [maplibre-gl-js' Docs](https://maplibre.org/maplibre-gl-js/docs/API/classes/Map/#setfeaturestate)
-   */
-  public setFeatureState(feature: Partial<FeatureIdentifier>, state: any) {
-    this.sourceDirective.setFeatureState(feature, state);
-  }
-
-  /**
-   * Removes the `state` of a feature, setting it back to the default behavior.
-   * 
-   * @param target Feature identifier. `source` may be omitted, will use this source id.
-   * @param key The key in the feature state to reset.
-   * 
-   * @see [maplibre-gl-js' Docs](https://maplibre.org/maplibre-gl-js/docs/API/classes/Map/#removeFeatureState)
-   */
-  removeFeatureState(target: Partial<FeatureIdentifier>, key?: string) {
-    this.sourceDirective.removeFeatureState(target, key);
-  }
-
-  /**
-   * Gets the `state` of a feature.
-   * 
-   * @param feature Feature identifier. `source` may be omitted, will use this source id.
-   * @returns The state of the feature: a set of key-value pairs that was assigned to the feature at runtime.
-   * 
-   * @see [maplibre-gl-js' Docs](https://maplibre.org/maplibre-gl-js/docs/API/classes/Map/#getFeatureState)
-   */
-  getFeatureState(feature: Partial<FeatureIdentifier>): any {
-    return this.sourceDirective.getFeatureState(feature);
-  }
-}
